Fully isolated command handler from rest of code
